Updated service implementations to return optional values + integrity checks

+++ src/main/java/paralaks_gmail_com/todolist/service/impl/TodoItemServiceImpl.java
@@ -1,11 +1,11 @@
 package paralaks_gmail_com.todolist.service.impl;
 
 import paralaks_gmail_com.todolist.model.TodoItem;
+import paralaks_gmail_com.todolist.model.TodoList;
 import paralaks_gmail_com.todolist.repository.TodoItemRepository;
 import paralaks_gmail_com.todolist.repository.TodoListRepository;
 import paralaks_gmail_com.todolist.service.TodoItemService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Example;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -23,16 +23,12 @@ public class TodoItemServiceImpl implements TodoItemService {
   }
 
   @Override
-  public List<TodoItem> list(Integer listId) {
-
+  public Optional<List<TodoItem>> list(Integer listId) {
     if (listId == null)
-      return todoItemRepository.findAll();
+      return Optional.of(todoItemRepository.findAll());
 
-    TodoItem todoItem = new TodoItem();
-    todoItem.setListId(listId);
-    Example<TodoItem> todoItemExample = Example.of(todoItem);
-
-    return todoItemRepository.findAll(todoItemExample);
+    Optional<TodoList> list = todoListRepository.findById(listId);
+    return list.map(TodoList::getItems).or(Optional::empty);
   }
 
   @Override
@@ -42,12 +38,19 @@ public class TodoItemServiceImpl implements TodoItemService {
 
   @Override
   public TodoItem addOne(TodoItem item) {
-    item.setTodoList(todoListRepository.findById(item.getListId()).orElse(null));
+    Optional<TodoList> list = todoListRepository.findById(item.getListId());
+    if (!list.isPresent())
+      throw new IllegalArgumentException("List not found");
+
+    item.setTodoList(list.get());
     return todoItemRepository.save(item);
   }
 
   @Override
   public String deleteOne(Integer id) {
+    if (!todoItemRepository.findById(id).isPresent())
+      return null;
+
     todoItemRepository.deleteById(id);
     return "{\"message\": \"Todo item deleted successfully.\"}";
   }
